feat: add personal data section
