+++ src/components/ProfileForm/ProfileForm.module.css
@@ -1,0 +1,14 @@
+.profile-form {
+  text-align: left;
+  min-height: 370px;
+  margin: 0 auto;
+  max-width: 500px;
+  width: 100%;
+  padding: 0 1rem 1.5rem;
+}
+
+.steps {
+  max-width: 900px;
+  margin: 1.5rem auto;
+  padding-left: 1rem;
+}
